Changed speed and acceleration to doubles.

+++ src/include/configure.h
@@ -1,8 +1,8 @@
 #define convertToMm(x) x * 25.4
 
 extern double stepRatio;
-extern int accelerationMax;
-extern int speedMax;
+extern double accelerationMax;
+extern double speedMax;
 extern int inchMeasurements;
 
 extern int configure( char *filename );
